ProyectoIntegrador3sem/hotel_reservations/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3        
certifi==2024.2.2        
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==5.0.6
django-rest-swagger==2.2.0
django-urls==1.1.3
djangorestframework==3.15.1
drf-yasg==1.21.7
filelock==3.13.3
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==21.2.0
h5py==3.11.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.0
keras==3.2.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mysqlclient==2.2.4
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
openapi-codec==1.3.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
platformdirs==4.2.0
protobuf==4.25.3
Pygments==2.17.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.2
wheel==0.43.0
whitenoise==6.5.0
wrapt==1.16.0
